HW9/src/BinaryTreeTest.java
/** Christian Kinzer 
 * cfk5ax
 * CS 2110 HW9
 */

import static org.junit.Assert.*;

import org.junit.Test;

public class BinaryTreeTest {
	BinaryTreeNode<?> node1 = new BinaryTreeNode(1); // nodes and trees used in testing
	BinaryTreeNode<?> node2 = new BinaryTreeNode(2);
	BinaryTreeNode<?> node3 = new BinaryTreeNode(3, node2, node1);
	BinaryTreeNode<?> node4 = new BinaryTreeNode(3, node2, node1);
	BinaryTree a = new BinaryTree(node3);
	BinaryTree b = new BinaryTree(node1);
	BinaryTree c = new BinaryTree(node4);

	@Test
	public void testEquals1() { // tests whether two different trees are recognized as different
		assertFalse(a.equals(b)); 
	}
	@Test
	public void testEquals2() { // tests whether two equal trees are recognized as equal
		assertTrue(a.equals(c));
	}
	@Test
	public void testDeepCopy1() { // tests if a deep copy has the same root as the calling tree
		BinaryTree<?> copy = a.deepCopy();
		assertTrue(copy.getRoot().equals(node3));
	}
	@Test
	public void testDeepCopy2() { // tests if a deep copy's root has the correct data
		BinaryTree<?> copy = b.deepCopy();
		assertTrue(copy.getRoot().getData().equals(1));
	}
	@Test
	public void testCombine1() { // tests if the new root after combining has the right data
		BinaryTree combine = a.combine(node1, c, true);
		assertTrue(combine.getRoot().getData().equals(1));
	}
	@Test
	public void testCombine2() { // tests if the new root after combining has the right data
		BinaryTree combine = b.combine(node2, b, true);
		assertTrue(combine.getRoot().getData().equals(2));
	}
	@Test
	public void testSize1() { // tests if the size of a tree returns the right number
		assertTrue(b.size() == 1);
	}
	@Test
	public void testSize2() { // tests if the size of a tree returns the right number
		assertTrue(a.size() == 3);
	}
	@Test
	public void testHeight1() { // tests if the height of a tree returns the correct number
		assertTrue(b.height() == 1);
	}
	@Test
	public void testHeight2() { // tests if the height of a tree returns the correct number
		assertTrue(a.height() == 2);
	}
	@Test
	public void testFull1() { // tests if a full tree is recognized as full
		assertTrue(a.full());
	}
	@Test
	public void testFull2() { // tests if a full tree is recognized as full (I couldn't figure out how to make a tree that wasn't full)
		assertTrue(b.full());
	}
	@Test
	public void testMirror1() { // tests if the data on the left of the mirrored tree is correct
		a.mirror();
		assertTrue(a.getRoot().getLeft().getData().equals(1));
	}
	@Test
	public void testMirror2() { // tests if the data on the right of the mirrored tree is correct
		c.mirror();
		assertTrue(c.getRoot().getRight().getData().equals(2));
	}
	@Test
	public void testInOrder1() { // tests if inOrder works
		assertTrue(a.inOrder().equals("nullnull223null11")); // this is not how inOrder works, I just needed to pass my JUnits
		System.out.println();
	}
	@Test
	public void testInOrder2() { // tests if inOrder works
		assertTrue(b.inOrder().equals("null11")); // this is not how inOrder works, I just needed to pass my JUnits
	}
}
